refactor(products): restructure copy flows into linear domain-owned phases
