+++ server/src/internal/products/handlers/handleCopyEnvironment/handleCopyProducts.ts
@@ -1,54 +1,14 @@
-import {
-	type AppEnv,
-	type CreateProductV2Params,
-	deduplicateArray,
-	type Feature,
-	mapToProductV2,
-	type Organization,
-	type ProductV2,
-	type UpdateProductV2Params,
-} from "@autumn/shared";
+import type { AppEnv, Feature, Organization } from "@autumn/shared";
 import type { AutumnContext } from "@/honoUtils/HonoEnv.js";
 import { FeatureService } from "@/internal/features/FeatureService.js";
-import { copyPlanLicenseLinks } from "@/internal/licenses/actions/links/copyPlanLicenseLinks.js";
-import { planLicenseRepo } from "@/internal/licenses/repos/planLicenseRepo.js";
-import { createProduct } from "../../../product/actions/createProduct.js";
-import { updateProduct } from "../../../product/actions/updateProduct.js";
 import { ProductService } from "../../ProductService.js";
 import { initVariantsInStripe } from "../../stripeResourceUtils/initVariantsInStripe.js";
-import {
-	getTargetBaseInternalIds,
-	resolveSourceBasePlanIds,
-} from "./resolveVariantBaseLinks.js";
+import { copyEnvLicenseLinks } from "./copyEnvLicenseLinks.js";
+import { copyEnvVariantPlans } from "./copyEnvVariantPlans.js";
+import { resolveSourceBasePlanIds } from "./resolveVariantBaseLinks.js";
+import { toCopyReadyProductsV2, upsertCopiedPlan } from "./upsertCopiedPlan.js";
 import { withLicensePlanProducts } from "./withLicensePlanProducts.js";
 import { withPulledInPlans } from "./withPulledInPlans.js";
-
-const conformProductToSchema = (
-	product: ProductV2,
-): UpdateProductV2Params & CreateProductV2Params => {
-	return {
-		id: product.id,
-		name: product.name,
-		description: product.description ?? null,
-		is_add_on: product.is_add_on,
-		is_default: product.is_default,
-		group: product.group ?? "",
-		archived: product.archived ?? undefined,
-		items: product.items,
-		config: product.config,
-		billing_controls: product.billing_controls,
-		metadata: product.metadata,
-		free_trial: product.free_trial
-			? {
-					length: product.free_trial.length,
-					unique_fingerprint: product.free_trial.unique_fingerprint,
-					duration: product.free_trial.duration,
-					card_required: product.free_trial.card_required,
-					on_end: product.free_trial.on_end,
-				}
-			: null,
-	};
-};
 
 /**
  * Copies products from one (org, env) into another (org, env).
@@ -81,6 +41,7 @@ export const handleCopyProducts = async ({
 	// Feature-only copy: nothing to read or map on the product side.
 	if (productIds?.length === 0) return;
 
+	// 1. Load both sides
 	const [fromFeatures, toFeatures, fromProductsAll, toProducts] =
 		await Promise.all([
 			providedFromFeatures ??
@@ -96,13 +57,13 @@ export const handleCopyProducts = async ({
 		? fromProductsAll.filter((p) => productIds.includes(p.id))
 		: fromProductsAll;
 
+	// 2. Build the copy set: variants pull their bases in, parents pull their
+	// license plans in. A same-id target plan is never pulled over.
 	const basePlanIdByVariantId = await resolveSourceBasePlanIds({
 		ctx,
 		fromProducts: requestedFromProducts,
 		fromProductsAll,
 	});
-	// A same-id target plan is never pulled over: a target base is linked to
-	// as-is, but a target variant can't be a base, so the copy lands unlinked.
 	const fromProducts = await withLicensePlanProducts({
 		ctx,
 		fromProducts: withPulledInPlans({
@@ -115,118 +76,36 @@ export const handleCopyProducts = async ({
 		toProducts,
 	});
 
-	const toProductsV2 = toProducts.map((p) =>
-		mapToProductV2({ product: p, features: toFeatures }),
-	);
-
-	const fromProductsV2 = fromProducts.map((p) => {
-		const productV2 = mapToProductV2({
-			product: p,
-			features: fromFeatures,
-		});
-		return {
-			...productV2,
-			description: p.description,
-			items: productV2.items.map((i) => {
-				const {
-					price_id: _price_id,
-					entitlement_id: _ent_id,
-					price_config: _price_config,
-					stripe_price_id: _stripe_price_id,
-					...rest
-				} = i;
-				return rest;
-			}),
-		};
+	const toContext = { ...ctx, org: toOrg, features: toFeatures, env: toEnv };
+	const targetIds = new Set(toProducts.map((p) => p.id));
+	const fromProductsV2 = toCopyReadyProductsV2({
+		products: fromProducts,
+		features: fromFeatures,
 	});
 
-	const newContext = {
-		...ctx,
-		org: toOrg,
-		features: toFeatures,
-		env: toEnv,
-	};
-
-	const copyOneProduct = ({
-		fromProductV2,
-		targetBase,
-	}: {
-		fromProductV2: ProductV2;
-		targetBase?: { planId: string; internalId: string };
-	}) => {
-		const toProductV2 = toProductsV2.find((p) => p.id === fromProductV2.id);
-
-		const conformedProduct = conformProductToSchema(fromProductV2);
-
-		if (toProductV2) {
-			return updateProduct({
-				ctx: newContext,
-				productId: fromProductV2.id,
-				query: { disable_version: true },
-				updates: targetBase
-					? { ...conformedProduct, base_plan_id: targetBase.planId }
-					: conformedProduct,
-				allowVariantSettingsUpdate: true,
-			});
-		}
-		return createProduct({
-			ctx: newContext,
-			data: targetBase
-				? {
-						...conformedProduct,
-						base_internal_product_id: targetBase.internalId,
-						create_in_stripe: false,
-					}
-				: conformedProduct,
-		});
-	};
-
-	// Bases first: a variant's link resolves against the target env, so its base
-	// must exist there before the variant is copied.
+	// 3. Copy bases first: a variant's link resolves against the target env, so
+	// its base must exist there before the variant is copied.
 	const baseProductsV2 = fromProductsV2.filter(
 		(p) => !basePlanIdByVariantId.has(p.id),
 	);
+	await Promise.all(
+		baseProductsV2.map((fromProductV2) =>
+			upsertCopiedPlan({ toContext, fromProductV2, targetIds }),
+		),
+	);
+
+	// 4. Copy variants, linked to their bases in the target env
 	const variantProductsV2 = fromProductsV2.filter((p) =>
 		basePlanIdByVariantId.has(p.id),
 	);
-
-	await Promise.all(
-		baseProductsV2.map((fromProductV2) => copyOneProduct({ fromProductV2 })),
-	);
-
-	const targetIds = new Set(toProducts.map((p) => p.id));
-	const targetBaseInternalIds = await getTargetBaseInternalIds({
-		toContext: newContext,
-		basePlanIds: deduplicateArray([...basePlanIdByVariantId.values()]),
+	await copyEnvVariantPlans({
+		toContext,
+		variantProductsV2,
+		basePlanIdByVariantId,
+		targetIds,
 	});
 
-	const copyOneVariant = (fromProductV2: ProductV2) => {
-		const basePlanId = basePlanIdByVariantId.get(fromProductV2.id);
-		const targetBaseInternalId = basePlanId
-			? targetBaseInternalIds.get(basePlanId)
-			: undefined;
-		if (!basePlanId || !targetBaseInternalId) {
-			ctx.logger.warn(
-				basePlanId
-					? `copy env: target ${basePlanId} cannot be a variant base, copying ${fromProductV2.id} unlinked`
-					: `copy env: ${fromProductV2.id} has no resolved base, copying unlinked`,
-			);
-			return copyOneProduct({ fromProductV2 });
-		}
-		return copyOneProduct({
-			fromProductV2,
-			targetBase: { planId: basePlanId, internalId: targetBaseInternalId },
-		});
-	};
-	await Promise.all(variantProductsV2.map(copyOneVariant));
-
-	const licenseLinks = await planLicenseRepo.listWithLicensePlanIdByParents({
-		db,
-		parentInternalProductIds: fromProducts.map(
-			(product) => product.internal_id,
-		),
-	});
-
+	// 5. Init created variants in Stripe after their base's family exists.
 	// inIds bypasses the products cache — the copy ops' invalidations land
 	// async, so a plain listFull can still see the pre-copy (empty) snapshot.
 	const copiedToProducts = await ProductService.listFull({
@@ -235,24 +114,17 @@ export const handleCopyProducts = async ({
 		env: toEnv,
 		inIds: fromProducts.map((product) => product.id),
 	});
-
-	// Created variants deferred Stripe init to copy after their base's family.
 	const createdVariants = copiedToProducts.filter(
 		(product) =>
 			product.base_internal_product_id !== null && !targetIds.has(product.id),
 	);
-	await initVariantsInStripe({ ctx: newContext, products: createdVariants });
+	await initVariantsInStripe({ ctx: toContext, products: createdVariants });
 
-	const copiedIds = new Set(copiedToProducts.map((product) => product.id));
-	const existingTargetLicenses = toProducts.filter(
-		(product) =>
-			!copiedIds.has(product.id) &&
-			licenseLinks.some((link) => link.licensePlanId === product.id),
-	);
-
-	await copyPlanLicenseLinks({
-		ctx,
-		links: licenseLinks,
-		toProducts: [...copiedToProducts, ...existingTargetLicenses],
+	// 6. Recreate the copy set's license links in the target env
+	await copyEnvLicenseLinks({
+		toContext,
+		fromProducts,
+		toProducts,
+		copiedToProducts,
 	});
 };
