+++ server/src/internal/products/handlers/handleCopyProduct/copyProductForOrgs.ts
@@ -1,10 +1,6 @@
 import {
 	type AppEnv,
-	type CreateFeature,
-	CreateFeatureSchema,
-	deduplicateArray,
 	ErrCode,
-	type Feature,
 	type Organization,
 	ProductAlreadyExistsError,
 } from "@autumn/shared";
@@ -12,34 +8,17 @@ import { invalidateProductsCache } from "@/external/redis/actions/productsCache/
 import type { AutumnContext } from "@/honoUtils/HonoEnv.js";
 import { addCreditSystemMeteredFeatureIds } from "@/internal/features/creditSystemUtils.js";
 import { FeatureService } from "@/internal/features/FeatureService.js";
-import { planLicenseRepo } from "@/internal/licenses/repos/planLicenseRepo.js";
 import { ProductService } from "@/internal/products/ProductService.js";
 import {
 	copyProduct,
 	initProductInStripe,
 } from "@/internal/products/productUtils.js";
 import RecaseError from "@/utils/errorUtils.js";
-import { generateId } from "@/utils/genUtils.js";
-import { copyBaseVariants, listSourceVariants } from "./copyBaseVariants.js";
+import { copyBaseVariants } from "./copyBaseVariants.js";
 import { copyLicenseLinksForPlanCopy } from "./copyLicenseLinksForPlanCopy.js";
+import { copyMissingFeatures } from "./copyMissingFeatures.js";
 import { listExistingTargetPlanIds } from "./copyPlanIntoTarget.js";
-
-const initNewFeature = ({
-	data,
-	orgId,
-	env,
-}: {
-	data: CreateFeature;
-	orgId: string;
-	env: AppEnv;
-}): Feature => ({
-	...data,
-	org_id: orgId,
-	env,
-	created_at: Date.now(),
-	internal_id: generateId("fe"),
-	archived: false,
-});
+import { loadSourcePlanFamily } from "./loadSourcePlanFamily.js";
 
 /**
  * Copies a single product between (org, env) pairs. When fromOrg === toOrg this
@@ -66,7 +45,7 @@ export const copyProductForOrgs = async ({
 	toId: string;
 	toName: string;
 }): Promise<void> => {
-	const { db, logger } = ctx;
+	const { db } = ctx;
 	const crossOrg = fromOrg.id !== toOrg.id;
 
 	if (!crossOrg && fromEnv === toEnv && fromProductId === toId) {
@@ -90,6 +69,7 @@ export const copyProductForOrgs = async ({
 		});
 	}
 
+	// 1. Load the source plan and both sides' features
 	const [fromFullProduct, fromFeatures, toFeatures] = await Promise.all([
 		ProductService.getFull({
 			db,
@@ -113,37 +93,9 @@ export const copyProductForOrgs = async ({
 		});
 	}
 
-	const variants = await listSourceVariants({
-		db,
-		base: fromFullProduct,
-		fromOrg,
-		fromEnv,
-	});
-
-	const sourceLicenseLinks =
-		await planLicenseRepo.listWithLicensePlanIdByParents({
-			db,
-			parentInternalProductIds: [
-				fromFullProduct.internal_id,
-				...variants.map((variant) => variant.internal_id),
-			],
-		});
-	const familyPlanIds = new Set([
-		fromFullProduct.id,
-		...variants.map((variant) => variant.id),
-	]);
-	const licensePlanIds = deduplicateArray(
-		sourceLicenseLinks.map((link) => link.licensePlanId),
-	).filter((planId) => !familyPlanIds.has(planId));
-	const sourceLicensePlans =
-		licensePlanIds.length > 0
-			? await ProductService.listFull({
-					db,
-					orgId: fromOrg.id,
-					env: fromEnv,
-					inIds: licensePlanIds,
-				})
-			: [];
+	// 2. Load the plan family: variants, license links, and license plans
+	const { variants, sourceLicenseLinks, sourceLicensePlans } =
+		await loadSourcePlanFamily({ db, base: fromFullProduct, fromOrg, fromEnv });
 
 	fromFullProduct.is_default = false;
 	if (crossOrg) {
@@ -152,16 +104,15 @@ export const copyProductForOrgs = async ({
 
 	const toContext = { ...ctx, org: toOrg, env: toEnv, features: toFeatures };
 
-	// A license plan already in the target is reused, not copied — pulling its
-	// features in anyway can abort the copy on unrelated type conflicts.
+	// 3. Copy the features the copied plans reference. A license plan already
+	// in the target is reused, not copied, so its features stay out of scope.
 	const existingTargetLicensePlanIds = await listExistingTargetPlanIds({
 		toContext,
-		planIds: licensePlanIds,
+		planIds: sourceLicensePlans.map((licensePlan) => licensePlan.id),
 	});
 	const licensePlansToCopy = sourceLicensePlans.filter(
 		(licensePlan) => !existingTargetLicensePlanIds.has(licensePlan.id),
 	);
-
 	const featureIdsToCopy = new Set(
 		[fromFullProduct, ...variants, ...licensePlansToCopy].flatMap((product) =>
 			product.entitlements.map((entitlement) => entitlement.feature.id),
@@ -171,36 +122,15 @@ export const copyProductForOrgs = async ({
 		features: fromFeatures,
 		featureIds: featureIdsToCopy,
 	});
-
 	if (crossOrg || fromEnv !== toEnv) {
-		for (const fromFeature of fromFeatures.filter((f) =>
-			featureIdsToCopy.has(f.id),
-		)) {
-			const toFeature = toFeatures.find((f) => f.id === fromFeature.id);
-
-			if (toFeature && fromFeature.type !== toFeature.type) {
-				throw new RecaseError({
-					message: `Feature ${fromFeature.name} exists in ${toEnv} with a different type. Please match them then try again.`,
-					code: ErrCode.InvalidRequest,
-					statusCode: 400,
-				});
-			}
-
-			if (!toFeature) {
-				const res = await FeatureService.insert({
-					db,
-					data: initNewFeature({
-						data: CreateFeatureSchema.parse(fromFeature),
-						orgId: toOrg.id,
-						env: toEnv,
-					}),
-					logger,
-				});
-				toFeatures.push(res![0]);
-			}
-		}
+		await copyMissingFeatures({
+			toContext,
+			fromFeatures,
+			featureIds: featureIdsToCopy,
+		});
 	}
 
+	// 4. Copy the base and init its Stripe resources
 	const toBaseInternalId = await copyProduct({
 		ctx: toContext,
 		product: fromFullProduct,
@@ -209,7 +139,6 @@ export const copyProductForOrgs = async ({
 		fromEnv,
 		fromFeatures,
 	});
-
 	const copiedBase = await ProductService.getFull({
 		db,
 		idOrInternalId: toBaseInternalId,
@@ -218,6 +147,7 @@ export const copyProductForOrgs = async ({
 	});
 	await initProductInStripe({ ctx: toContext, product: copiedBase });
 
+	// 5. Copy the base's variants, relinked to the copied base
 	const copiedVariantIds = await copyBaseVariants({
 		toContext,
 		variants,
@@ -227,6 +157,7 @@ export const copyProductForOrgs = async ({
 		crossOrg,
 	});
 
+	// 6. Recreate the family's license links in the target
 	await copyLicenseLinksForPlanCopy({
 		toContext,
 		fromBase: fromFullProduct,
